pyproject.toml
[tool.poetry]
name = "faust_consumers"
version = "0.1.0"
description = ""
authors = ["Christophoros <bekos.christopher@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
python-dotenv = "^0.20.0"
faust-streaming = "^0.8.6"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
